+++ PXCTestKit/CommandLineInterface.swift
@@ -26,13 +26,13 @@ import Foundation
                        Option<ExistingFileURL>("deviceset", ExistingFileURL(url: URL(fileURLWithPath: FBSimulatorControlConfiguration.defaultDeviceSetPath())), description: "Path to the Simulator device set."),
                        Option<FileURL>("output", FileURL(url: URL(fileURLWithPath: "output")), description: "Output path where the test reports and log files should be stored."),
                        Option<Locale>("locale", Locale(identifier: "en"), description: "Locale to set for the Simulator."),
-                       Option<Preferences>("preferences", Preferences(), description: "Path to some preferences.json to be applied with the Simulator."),
+                       Option<Defaults>("defaults", Defaults(), description: "Path to some defaults.json to be applied with the Simulator."),
                        Option<Reporter>("reporter", .rspec, description: "Console reporter type. Supported values: rspec, json"),
                        VaradicOption<Only>("only", [], description: "Comma separated list of tests that should be executed only. Format: TARGET[:Class/case[,Class2/case2]]"),
                        VaradicOption<Destination>("destination", [Destination()], description: "A comma-separated set of key=value pairs describing the destination to use. Default: \(Destination().description)"),
                        Option<Double>("timeout", 3600.0, description: "Timeout in seconds for the test execution to finish."),
                        Flag("no-color", description: "Do not add colors to console output.")
-            ) { (testRun, deviceSet, output, locale, preferences, reporter, only, destination, timeout, noColor) in
+            ) { (testRun, deviceSet, output, locale, defaults, reporter, only, destination, timeout, noColor) in
                 ANSI.disabled = noColor
                 let consoleOutput = ConsoleOutput()
 
@@ -42,7 +42,7 @@ import Foundation
                     output: OutputManager(url: output.url),
                     locale: locale,
                     environment: ProcessInfo.processInfo.environment,
-                    preferences: preferences.dictionary,
+                    defaults: defaults.dictionary,
                     reporterType: reporter.type,
                     testsToRun: only.dictionary(),
                     simulatorConfigurations: destination.map({ $0.simulatorConfiguration }),