+++ PXCTestKit/Extension/FBSimulatorInteraction+PXCTestKit.swift
@@ -11,13 +11,13 @@ import Foundation
 
 extension FBSimulatorInteraction {
 
-    func loadPreferences(_ defaults: [String: Any]) -> FBSimulatorInteraction {
+    func loadDefaults(inDomainOrPath domainOrPath: String?, defaults: [String : Any]) -> FBSimulatorInteraction {
         return interact(simulator: { (error: NSErrorPointer, simulator: FBSimulator?) -> Bool in
             guard let simulator = simulator else { return false }
             let strategy = FBDefaultsModificationStrategy(simulator: simulator)
             var result = true
             do {
-                try strategy.modifyDefaults(inDomainOrPath: "com.apple.Preferences", defaults: defaults)
+                try strategy.modifyDefaults(inDomainOrPath: domainOrPath, defaults: defaults)
             }
             catch (let innerError) {
                 error?.pointee = innerError as NSError