+++ PXCTestKit/Option/Defaults.swift
@@ -1,5 +1,5 @@
 //
-//  Preferences.swift
+//  Defaults.swift
 //  pxctest
 //
 //  Created by Johannes Plunien on 24/11/16.
@@ -11,34 +11,35 @@ import Foundation
 
 /*
  * Loads some JSON from a file containing a dictionary of optians that will
- * be written to a Simulator's Library/Preferences/com.apple.Preferences.plist
- * property list file. Example:
+ * be written to a Simulator's defaults. Example:
  *
  * {
- *   "KeyboardAllowPaddle": false,
- *   "KeyboardAssistant": false,
- *   "KeyboardAutocapitalization": false,
- *   "KeyboardAutocorrection": false,
- *   "KeyboardCapsLock": false,
- *   "KeyboardCheckSpelling": false,
- *   "KeyboardPeriodShortcut": false,
- *   "KeyboardPrediction": false,
- *   "KeyboardShowPredictionBar": false
+ *   "com.apple.Preferences": {
+ *     "KeyboardAllowPaddle": false,
+ *     "KeyboardAssistant": false,
+ *     "KeyboardAutocapitalization": false,
+ *     "KeyboardAutocorrection": false,
+ *     "KeyboardCapsLock": false,
+ *     "KeyboardCheckSpelling": false,
+ *     "KeyboardPeriodShortcut": false,
+ *     "KeyboardPrediction": false,
+ *     "KeyboardShowPredictionBar": false
+ *   }
  * }
  *
  * This content will disable all the on/off keyboard helpers you can find in the
  * Simulator's Settings app.
  */
 
-struct Preferences: ArgumentConvertible {
+struct Defaults: ArgumentConvertible {
 
-    let dictionary: [String: Any]
+    let dictionary: [String: [String: Any]]
 
     var description: String {
         return dictionary.description
     }
 
-    init(dictionary: [String: Any]) {
+    init(dictionary: [String: [String: Any]]) {
         self.dictionary = dictionary
     }
 
@@ -49,11 +50,11 @@ struct Preferences: ArgumentConvertible {
     init(parser: ArgumentParser) throws {
         if let value = parser.shift() {
             let data = try Data(contentsOf: URL(fileURLWithPath: value))
-            if let dictionary = try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any] {
+            if let dictionary = try JSONSerialization.jsonObject(with: data, options: []) as? [String: [String: Any]] {
                 self.init(dictionary: dictionary)
             }
             else {
-                throw ArgumentError.invalidType(value: value, type: "preferences", argument: nil)
+                throw ArgumentError.invalidType(value: value, type: "defaults", argument: nil)
             }
         } else {
             throw ArgumentError.missingValue(argument: nil)